fix: edge case causing time to be 1 day ahead of what it should be
The docker container running the tasks are in UTC, and when pendulum
uses this to infer dates, it will be UTC, which will cause anything that
is past 7:00 PM EST  be considered to be the next day.

This fixes this issue by first anchoring the date in the designated time
zone of the hospital before setting the time.